requirements.txt
# Core
Flask>=3.0,<4.0
Flask-SQLAlchemy>=3.1,<4.0
Flask-Migrate>=4.0,<5.0
Flask-CORS>=5.0,<6.0

# Serialization & Validation
marshmallow>=3.20,<4.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment
python-dotenv>=1.0,<2.0

# Production server
gunicorn>=22.0,<24.0

# Testing
pytest>=8.0,<9.0
pytest-cov>=5.0,<7.0
